Backend/Ecatalogue/requirements.txt
arrow==1.3.0
asgiref==3.8.1
attrs==23.2.0
binaryornot==0.4.4
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==5.0.4
django-allauth==0.61.1
django-appconf==1.0.6
django-browser-reload==1.12.1
django-compressor==4.4
django-cors-headers==4.3.1
django-filter==24.2
django-rest-swagger==2.2.0
django-tailwind==3.8.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
oauthlib==3.2.2
odict==1.9.0
openapi-codec==1.3.2
packaging==24.0
pillow==10.3.0
plumber==1.7
psycopg2==2.9.9
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
rcssmin==1.1.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
rich==13.7.1
rjsmin==1.2.1
rpds-py==0.18.0
setuptools==69.5.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.event==5.0
zope.hookable==6.0
zope.interface==6.3
zope.lifecycleevent==5.0
zope.proxy==5.2
